version 4

+++ assign4.pde
@@ -127,8 +127,8 @@ void keyPressed() {
 /*---------Make Alien Function-------------*/
 void alienMaker(float alienNum, float inputCol) {
   for (int i =0; i<alienNum; i++) {
-    float col = i % inputCol;
-    float row = i / inputCol;
+    int col = i % inputCol;
+    int row = int(i / inputCol);
     aList[i] = new Alien(50 + col*40, 50 + row*50);
   }
 }
